pip_requirements.txt
Flask==0.10.1
newman==1.0.0
PyChef==0.2.2
pyrax==1.5.0
sqlcontext==0.0.1
mock>=1.0
PyYAML==3.10
